requirements.txt
streamlit 
simplejson 
openai 
argparse 
loguru 
pdf2docx 
docx
pyYaml